fix: axios-retry configuration
