+++ src/rotacloud.ts
@@ -1,4 +1,5 @@
 import axios, { AxiosError, isAxiosError } from 'axios';
+import axiosRetry, { isNetworkOrIdempotentRequestError } from 'axios-retry';
 import {
   AccountsService,
   AttendanceService,
@@ -14,7 +15,6 @@ import {
   LeaveTypesService,
   LeaveService,
   LocationsService,
-  RetryStrategy,
   RolesService,
   SettingsService,
   ShiftsService,
@@ -26,9 +26,24 @@ import {
   UsersService,
   UsersClockInService,
 } from './services/index.js';
-import { SDKBase, SDKConfig } from './interfaces/index.js';
+import { RetryOptions, RetryStrategy, SDKBase, SDKConfig } from './interfaces/index.js';
 import { PinsService } from './services/pins.service.js';
 import { SDKError } from './models/index.js';
+
+const DEFAULT_RETRIES = 3;
+const DEFAULT_RETRY_DELAY = 2000;
+
+const DEFAULT_RETRY_STRATEGY_OPTIONS: Record<RetryStrategy, RetryOptions> = {
+  [RetryStrategy.Exponential]: {
+    exponential: true,
+    maxRetries: DEFAULT_RETRIES,
+  },
+  [RetryStrategy.Static]: {
+    exponential: false,
+    maxRetries: DEFAULT_RETRIES,
+    delay: DEFAULT_RETRY_DELAY,
+  },
+};
 
 const DEFAULT_CONFIG: Partial<SDKBase> = {
   baseUri: 'https://api.rotacloud.com/v1',
@@ -89,8 +104,31 @@ export class RotaCloud {
 
   set config(configVal: SDKConfig) {
     RotaCloud.config = configVal;
+    this.setupInterceptors(configVal.retry, configVal.interceptors);
+  }
+
+  private setupInterceptors(retry: SDKConfig['retry'], customInterceptors: SDKConfig['interceptors']) {
     this.client.interceptors.request.clear();
     this.client.interceptors.response.clear();
+
+    // NOTE: Retry interceptor - must be setup first
+    if (retry) {
+      const retryConfig = typeof retry === 'string' ? DEFAULT_RETRY_STRATEGY_OPTIONS[retry] : retry;
+
+      axiosRetry(this.client, {
+        retries: retryConfig.maxRetries,
+        shouldResetTimeout: true,
+        retryCondition: (err: AxiosError) => isNetworkOrIdempotentRequestError(err) || err.response?.status === 429,
+        retryDelay: (retryCount) => {
+          if (retryConfig.exponential) {
+            return axiosRetry.exponentialDelay(retryCount);
+          }
+          return retryConfig.delay;
+        },
+      });
+    }
+
+    // NOTE: Error interceptor - must be setup after retry but before custom
     this.client.interceptors.response.use(
       (response) => response,
       (error: unknown) => {
@@ -101,14 +139,16 @@ export class RotaCloud {
         return Promise.reject(parsedError);
       },
     );
-    for (const requestInterceptor of this.config.interceptors?.request ?? []) {
+
+    // NOTE: Custom interceptors - must be setup last
+    for (const requestInterceptor of customInterceptors?.request ?? []) {
       const callbacks =
         typeof requestInterceptor === 'function'
           ? ([requestInterceptor] as const)
           : ([requestInterceptor?.success, requestInterceptor?.error] as const);
       this.client.interceptors.request.use(...callbacks);
     }
-    for (const responseInterceptor of this.config.interceptors?.response ?? []) {
+    for (const responseInterceptor of customInterceptors?.response ?? []) {
       const callbacks =
         typeof responseInterceptor === 'function'
           ? ([responseInterceptor] as const)
@@ -118,7 +158,7 @@ export class RotaCloud {
   }
 }
 
-export { RetryStrategy, RetryOptions } from './services/service.js';
+export { RetryStrategy, RetryOptions };
 export * from './interfaces/index.js';
 export * from './interfaces/query-params/index.js';
 export * from './models/index.js';
